portal/requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.4
django-import-export==4.0.0
django-jazzmin==3.0.0
pillow==10.3.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
